+++ EssentialApp/EssentialApp/FeedUIComposer.swift
@@ -14,8 +14,8 @@ public final class FeedUIComposer {
     private init() {}
     
     public static func feedComposedWith(feedLoader: @escaping () -> AnyPublisher<[FeedImage], Error>, imageLoader: @escaping (URL) -> FeedImageDataLoader.Publisher) -> FeedViewController {
-            let presentationAdapter = FeedLoaderPresentationAdapter(
-                feedLoader: { feedLoader().dispatchOnMainQueue() })
+            let presentationAdapter = LoadResourcePresentationAdapter<[FeedImage], FeedViewAdapter>(
+                loader: { feedLoader().dispatchOnMainQueue() })
         
         let feedController = makeFeedViewController(
             delegate: presentationAdapter,
